requirements.txt
manga_ocr
numpy>=1.21.2,<2.1.0
opencv-python>=4.10.0,<5.0.0
pillow>=10.0.0,<11.0.0
